feat(optimize memory use): can set "needEdge" to false, if don't call edge()
if you don't call function edge(), you don't need gTXOMap. disable gTXOMap will save a lot of memory.
